requirements.txt
PySide6==6.7.2
Pillow==10.4.0